uodate cmake

+++ Project/main.cpp
@@ -50,7 +50,7 @@ int main()
     printLocalResults("lresults.txt",dfn.Fractures,dfn.Traces);
 
     CreateMesh(dfn.Fractures,tol,dfn,plm);
-    for(int i=0; i<plm.meshes.size(); i++){
+    for(size_t i=0; i<plm.meshes.size(); i++){
         cout << "Numero di celle 2D: " << plm.meshes[i].numPolygons << endl;
     }
